Move the TextAndIconButtonStyle to its component
