+++ src/modules/UI/components/Buttons/TextAndIconButton.ui.js
@@ -4,6 +4,7 @@ import React, { Component } from 'react'
 import { StyleSheet, Text, TouchableHighlight, View } from 'react-native'
 import FAIcon from 'react-native-vector-icons/MaterialIcons'
 
+import { THEME } from '../../../../theme/variables/airbitz.js'
 import { styles } from './style'
 
 type Props = {
@@ -16,7 +17,7 @@ type State = {
   pressed: boolean
 }
 
-class TextAndIconButton extends Component<Props, State> {
+export class TextAndIconButton extends Component<Props, State> {
   UNSAFE_componentWillMount() {
     this.setState({
       pressed: false
@@ -75,4 +76,45 @@ class TextAndIconButton extends Component<Props, State> {
   }
 }
 
-export { TextAndIconButton }
+export const TextAndIconButtonStyle = {
+  container: {
+    width: '100%',
+    height: '100%',
+    justifyContent: 'space-around',
+    alignItems: 'center'
+  },
+  centeredContent: {
+    width: '100%',
+    alignItems: 'center'
+  },
+  inner: {
+    position: 'relative',
+    flexDirection: 'row'
+  },
+  textContainer: {
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'space-around'
+  },
+  iconContainer: {
+    height: '100%',
+    flexDirection: 'column',
+    alignItems: 'center'
+  },
+  text: {
+    color: THEME.COLORS.WHITE,
+    fontSize: 20
+  },
+  textPressed: {
+    color: THEME.COLORS.GRAY_2,
+    fontSize: 20
+  },
+  icon: {
+    color: THEME.COLORS.WHITE
+  },
+  iconPressed: {
+    color: THEME.COLORS.GRAY_2
+  },
+  iconSize: 25,
+  underlayColor: THEME.COLORS.TRANSPARENT
+}
