+++ src/styles/scenes/CryptoExchangeSceneStyles.js
@@ -1,8 +1,8 @@
 // @flow
 
+import { TextAndIconButtonStyle } from '../../modules/UI/components/Buttons/TextAndIconButton.ui.js'
 import THEME from '../../theme/variables/airbitz'
 import { scale } from '../../util/scaling.js'
-import { TextAndIconButtonStyle } from '../components/ButtonStyles.js'
 
 export const styles = {
   mainScrollView: {
